Single redactor class of classes.
